Recommender for Ecommerce Wesite/backend/requirements.txt
Django>=4.2
djangorestframework>=3.14.0
python-dotenv>=1.0.0
mysqlclient>=2.1.0
django-cors-headers>=4.0.0
drf-spectacular>=0.26.0
drf-spectacular-sidecar>=0.12.0
